Refactored post processing for better supporting multiple post processors
