+++ tests/mock_failing_oidn.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+exit 1
